+++ Build.ps1
@@ -2,6 +2,7 @@ echo "In directory: $PSScriptRoot"
 
 $solution = "serilog-sinks-elasticsearch.sln"
 $test = "test\\Serilog.Sinks.Elasticsearch.Tests\\Serilog.Sinks.Elasticsearch.Tests.csproj"
+$testIntegration = "test\\Serilog.Sinks.Elasticsearch.IntegrationTests\\Serilog.Sinks.Elasticsearch.IntegrationTests.csproj"
 [string[]]$projects = @(
     ("src\\Serilog.Sinks.Elasticsearch\\Serilog.Sinks.Elasticsearch.csproj"),
     ("src\\Serilog.Formatting.Elasticsearch\\Serilog.Formatting.Elasticsearch.csproj")
@@ -20,6 +21,16 @@ function Invoke-Build()
     if($LASTEXITCODE -ne 0) 
     {
         Write-Output "The tests failed"
+        exit 1 
+    }
+    
+    Write-Output "Running integration tests"
+    # Tee-Object forces console redirection on vstest which magically makes Console.WriteLine works again.
+    # This allows you to see the console out of Elastic.Xunit while its running
+    & dotnet test $testIntegration -c Release | Tee-Object -Variable integ 
+    if($LASTEXITCODE -ne 0) 
+    {
+        Write-Output "The integration tests failed"
         exit 1 
     }