+++ src/Serilog.Sinks.Elasticsearch/Sinks/ElasticSearch/ElasticSearchTemplateProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using Elasticsearch.Net.Specification.IndicesApi;
 
 namespace Serilog.Sinks.Elasticsearch
 {
@@ -19,7 +20,11 @@ namespace Serilog.Sinks.Elasticsearch
         /// <summary>
         /// Elasticsearch version &gt;= version 6.0
         /// </summary>
-        ESv6 = 2
+        ESv6 = 2,
+        /// <summary>
+        /// Elasticsearch version &gt;= version 7.0
+        /// </summary>
+        ESv7 = 3
     }
 
     /// <summary>
@@ -27,13 +32,7 @@ namespace Serilog.Sinks.Elasticsearch
     /// </summary>
     public class ElasticsearchTemplateProvider
     {
-        ///  <summary>
-        ///
-        ///  </summary>
-        ///  <param name="settings"></param>
-        /// <param name="templateMatchString"></param>
-        /// <param name="version"></param>
-        ///  <returns></returns>
+        [Obsolete("Use the overload taking ElasticsearchSinkOptions which takes IncludeTypeName into account")]
         public static object GetTemplate(
             Dictionary<string, string> settings,
             string templateMatchString,
@@ -44,97 +43,112 @@ namespace Serilog.Sinks.Elasticsearch
                 case AutoRegisterTemplateVersion.ESv5:
                     return GetTemplateESv5(settings, templateMatchString);
                 case AutoRegisterTemplateVersion.ESv6:
-                    return GetTemplateESv6(settings, templateMatchString);
+                    return GetTemplateESv6(null, null, settings, templateMatchString);
+                case AutoRegisterTemplateVersion.ESv7:
+                    return GetTemplateESv7(null, null, settings, templateMatchString);
                 case AutoRegisterTemplateVersion.ESv2:
                     return GetTemplateESv2(settings, templateMatchString);
                 default:
                     throw new ArgumentOutOfRangeException(nameof(version), version, null);
             }
         }
+        
+        public static object GetTemplate(ElasticsearchSinkOptions options,
+            string discoveredVersion,
+            Dictionary<string, string> settings,
+            string templateMatchString,
+            AutoRegisterTemplateVersion version = AutoRegisterTemplateVersion.ESv2)
+        {
+            switch (version)
+            {
+                case AutoRegisterTemplateVersion.ESv5:
+                    return GetTemplateESv5(settings, templateMatchString);
+                case AutoRegisterTemplateVersion.ESv2:
+                    return GetTemplateESv2(settings, templateMatchString);
+                case AutoRegisterTemplateVersion.ESv6:
+                    return GetTemplateESv6(options, discoveredVersion, settings, templateMatchString);
+                case AutoRegisterTemplateVersion.ESv7:
+                    return GetTemplateESv7(options, discoveredVersion, settings, templateMatchString);
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(version), version, null);
+            }
+        }
 
-        private static object GetTemplateESv6(
+        private static object GetTemplateESv7(ElasticsearchSinkOptions options, string discoveredVersion,
             Dictionary<string, string> settings,
             string templateMatchString)
         {
-            return new
+            object mappings = new
             {
-                index_patterns = new[] { templateMatchString },
-                settings = settings,
-                mappings = new
+                dynamic_templates = new List<Object>
                 {
-                    _default_ = new
+                    //when you use serilog as an adaptor for third party frameworks
+                    //where you have no control over the log message they typically
+                    //contain {0} ad infinitum, we force numeric property names to
+                    //contain strings by default.
                     {
-                        dynamic_templates = new List<Object>
+                        new
                         {
-                            //when you use serilog as an adaptor for third party frameworks
-                            //where you have no control over the log message they typically
-                            //contain {0} ad infinitum, we force numeric property names to
-                            //contain strings by default.
+                            numerics_in_fields = new
                             {
-                                new
+                                path_match = @"fields\.[\d+]$",
+                                match_pattern = "regex",
+                                mapping = new
                                 {
-                                    numerics_in_fields = new
-                                    {
-                                        path_match = @"fields\.[\d+]$",
-                                        match_pattern = "regex",
-                                        mapping = new
-                                        {
-                                            type = "text",
-                                            index = true,
-                                            norms = false
-                                        }
-                                    }
+                                    type = "text",
+                                    index = true,
+                                    norms = false
                                 }
-                            },
+                            }
+                        }
+                    },
+                    {
+                        new
+                        {
+                            string_fields = new
                             {
-                                new
+                                match = "*",
+                                match_mapping_type = "string",
+                                mapping = new
                                 {
-                                    string_fields = new
+                                    type = "text",
+                                    index = true,
+                                    norms = false,
+                                    fields = new
                                     {
-                                        match = "*",
-                                        match_mapping_type = "string",
-                                        mapping = new
+                                        raw = new
                                         {
-                                            type = "text",
+                                            type = "keyword",
                                             index = true,
-                                            norms = false,
-                                            fields = new
-                                            {
-                                                raw = new
-                                                {
-                                                    type = "keyword",
-                                                    index = true,
-                                                    ignore_above = 256
-                                                }
-                                            }
+                                            ignore_above = 256
                                         }
                                     }
                                 }
                             }
-                        },
-                        properties = new Dictionary<string, object>
+                        }
+                    }
+                },
+                properties = new Dictionary<string, object>
+                {
+                    {"message", new {type = "text", index = true}},
+                    {
+                        "exceptions", new
                         {
-                            {"message", new {type = "text", index = "true"}},
+                            type = "nested",
+                            properties = new Dictionary<string, object>
                             {
-                                "exceptions", new
+                                {"Depth", new {type = "integer"}},
+                                {"RemoteStackIndex", new {type = "integer"}},
+                                {"HResult", new {type = "integer"}},
+                                {"StackTraceString", new {type = "text", index = true}},
+                                {"RemoteStackTraceString", new {type = "text", index = true}},
                                 {
-                                    type = "nested",
-                                    properties = new Dictionary<string, object>
+                                    "ExceptionMessage", new
                                     {
-                                        {"Depth", new {type = "integer"}},
-                                        {"RemoteStackIndex", new {type = "integer"}},
-                                        {"HResult", new {type = "integer"}},
-                                        {"StackTraceString", new {type = "text", index = "true"}},
-                                        {"RemoteStackTraceString", new {type = "text", index = "true"}},
+                                        type = "object",
+                                        properties = new Dictionary<string, object>
                                         {
-                                            "ExceptionMessage", new
-                                            {
-                                                type = "object",
-                                                properties = new Dictionary<string, object>
-                                                {
-                                                    {"MemberType", new {type = "integer"}},
-                                                }
-                                            }
+                                            {"MemberType", new {type = "integer"}},
                                         }
                                     }
                                 }
@@ -142,6 +156,108 @@ namespace Serilog.Sinks.Elasticsearch
                         }
                     }
                 }
+            };
+            mappings = discoveredVersion?.StartsWith("6.") ?? false ? new { _doc = mappings } : mappings;
+
+            return new
+            {
+                index_patterns = new[] {templateMatchString},
+                settings = settings,
+                mappings = mappings
+            };
+        }
+
+        private static object GetTemplateESv6(ElasticsearchSinkOptions options, string discoveredVersion,
+            Dictionary<string, string> settings,
+            string templateMatchString)
+        {
+            object mappings = new
+            {
+                dynamic_templates = new List<Object>
+                {
+                    //when you use serilog as an adaptor for third party frameworks
+                    //where you have no control over the log message they typically
+                    //contain {0} ad infinitum, we force numeric property names to
+                    //contain strings by default.
+                    {
+                        new
+                        {
+                            numerics_in_fields = new
+                            {
+                                path_match = @"fields\.[\d+]$",
+                                match_pattern = "regex",
+                                mapping = new
+                                {
+                                    type = "text",
+                                    index = true,
+                                    norms = false
+                                }
+                            }
+                        }
+                    },
+                    {
+                        new
+                        {
+                            string_fields = new
+                            {
+                                match = "*",
+                                match_mapping_type = "string",
+                                mapping = new
+                                {
+                                    type = "text",
+                                    index = true,
+                                    norms = false,
+                                    fields = new
+                                    {
+                                        raw = new
+                                        {
+                                            type = "keyword",
+                                            index = true,
+                                            ignore_above = 256
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                },
+                properties = new Dictionary<string, object>
+                {
+                    {"message", new {type = "text", index = true}},
+                    {
+                        "exceptions", new
+                        {
+                            type = "nested",
+                            properties = new Dictionary<string, object>
+                            {
+                                {"Depth", new {type = "integer"}},
+                                {"RemoteStackIndex", new {type = "integer"}},
+                                {"HResult", new {type = "integer"}},
+                                {"StackTraceString", new {type = "text", index = true}},
+                                {"RemoteStackTraceString", new {type = "text", index = true}},
+                                {
+                                    "ExceptionMessage", new
+                                    {
+                                        type = "object",
+                                        properties = new Dictionary<string, object>
+                                        {
+                                            {"MemberType", new {type = "integer"}},
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            };
+            
+            mappings = discoveredVersion?.StartsWith("7.") ?? false ? (object) new { _doc = mappings} : new { _default_ = mappings};
+            
+            return new
+            {
+                index_patterns = new[] { templateMatchString },
+                settings = settings,
+                mappings = mappings
             };
         }